build: split buildflags package
Planned to be imported by nerdctl in future.
